fix(MnemonicPhrase): Move component to component file
